Move user_feed and explore out of the common
