--- lib/assets/javascripts/cartodb/common/views/explore/dropdown_template.jst.ejs
@@ -1,7 +1,0 @@
-<ul class="SettingsDropdown">
-  <li class="SettingsDropdown-item">
-    <p><a class="js-both" href="#/both">Maps and datasets</a></p>
-    <p><a class="js-maps" href="#/maps">Maps</a></p>
-    <p><a class="js-datasets" href="#/datasets">Datasets</a></p>
-  </li>
-</ul>
