+++ lib/assets/javascripts/cartodb/explore/dataset_item_template.jst.ejs
