+++ lib/assets/javascripts/cartodb/explore/dropdown.js
@@ -22,7 +22,7 @@ module.exports = cdb.admin.DropdownMenu.extend({
 
   initialize: function() {
     this.elder('initialize');
-    this.template_base = cdb.templates.getTemplate('common/views/explore/dropdown_template');
+    this.template_base = cdb.templates.getTemplate('explore/dropdown_template');
 
     // Necessary to hide dialog on click outside popup, for example.
     cdb.god.bind('closeDialogs', this.hide, this);
