+++ lib/assets/javascripts/cartodb/explore/dropdown_template.jst.ejs
@@ -1,0 +1,7 @@
+<ul class="SettingsDropdown">
+  <li class="SettingsDropdown-item">
+    <p><button class="SettingsDropdown-itemLink js-both">Maps and datasets</button></p>
+    <p><button class="SettingsDropdown-itemLink js-maps">Maps</button></p>
+    <p><button class="SettingsDropdown-itemLink js-datasets">Datasets</button></p>
+  </li>
+</ul>
