+++ lib/assets/javascripts/cartodb/explore/entry.js
@@ -5,8 +5,8 @@ var UserTourView = require('../public_common/user_tour_view');
 var UserIndustriesView = require('../public_common/user_industries_view');
 var UserResourcesView = require('../public_common/user_resources_view');
 var FavMapView = require('../public_dashboard/fav_map_view');
-var Explore = require('../common/views/explore/view');
-var ExploreModel = require('../common/views/explore/model');
+var Explore = require('./view');
+var ExploreModel = require('./model');
 
 /**
  * Entry point for the user feed, bootstraps all dependency models and application.
