+++ lib/assets/javascripts/cartodb/explore/feed_collection.js
