+++ lib/assets/javascripts/cartodb/explore/loading.jst.ejs
