+++ lib/assets/javascripts/cartodb/explore/map_item_template.jst.ejs
