+++ lib/assets/javascripts/cartodb/explore/model.js
