+++ lib/assets/javascripts/cartodb/explore/template.jst.ejs
