+++ lib/assets/javascripts/cartodb/explore/view.js
@@ -1,10 +1,9 @@
 var cdb = require('cartodb.js');
-var MapCardPreview = require('../mapcard_preview');
-var ScrollableHeader = require('../scrollable_header');
+var MapCardPreview = require('../common/views/mapcard_preview');
+var ScrollableHeader = require('../common/views/scrollable_header');
 var VisualizationDropdown = require('./dropdown');
 var Visualizations = require('./feed_collection');
-
-var Feed = require('../feed/view');
+var Feed = require('../user_feed/view');
 
 module.exports = Feed.extend({
   tagName: 'div',
@@ -92,10 +91,10 @@ module.exports = Feed.extend({
   },
 
   _initTemplates: function() {
-    this.template = cdb.templates.getTemplate('common/views/explore/template');
-    this.mapTemplate = cdb.templates.getTemplate('common/views/explore/map_item_template');
-    this.datasetTemplate = cdb.templates.getTemplate('common/views/explore/dataset_item_template');
-    this.loaderTemplatePath = 'common/views/explore/loading';
+    this.template = cdb.templates.getTemplate('explore/template');
+    this.mapTemplate = cdb.templates.getTemplate('explore/map_item_template');
+    this.datasetTemplate = cdb.templates.getTemplate('explore/dataset_item_template');
+    this.loaderTemplatePath = 'explore/loading';
   },
 
   _initModels: function() {
