+++ lib/assets/javascripts/cartodb/user_feed/dataset_item_template.jst.ejs
