+++ lib/assets/javascripts/cartodb/user_feed/empty.jst.ejs
