+++ lib/assets/javascripts/cartodb/user_feed/entry.js
@@ -5,7 +5,7 @@ var UserTourView = require('../public_common/user_tour_view');
 var UserIndustriesView = require('../public_common/user_industries_view');
 var UserResourcesView = require('../public_common/user_resources_view');
 var FavMapView = require('../public_dashboard/fav_map_view');
-var Feed = require('../common/views/feed/view');
+var Feed = require('./view');
 var ScrollableHeader = require('../common/views/scrollable_header');
 
 /**
