+++ lib/assets/javascripts/cartodb/user_feed/feed_collection.js
