+++ lib/assets/javascripts/cartodb/user_feed/loading.jst.ejs
