+++ lib/assets/javascripts/cartodb/user_feed/map_item_template.jst.ejs
