+++ lib/assets/javascripts/cartodb/user_feed/template.jst.ejs
