+++ lib/assets/javascripts/cartodb/user_feed/view.js
@@ -1,8 +1,8 @@
 var cdb = require('cartodb.js');
-var ViewFactory = require('../../view_factory');
-var randomQuote = require('../../view_helpers/random_quote');
-var LikeView = require('../likes/view');
-var MapCardPreview = require('../mapcard_preview');
+var ViewFactory = require('../common/view_factory');
+var randomQuote = require('../common/view_helpers/random_quote');
+var LikeView = require('../common/views/likes/view');
+var MapCardPreview = require('../common/views/mapcard_preview');
 var Visualizations = require('./feed_collection');
 
 module.exports = cdb.core.View.extend({
@@ -49,12 +49,12 @@ module.exports = cdb.core.View.extend({
   },
 
   _initTemplates: function() {
-    this.template = cdb.templates.getTemplate('common/views/feed/template');
-    this.mapTemplate = cdb.templates.getTemplate('common/views/feed/map_item_template');
-    this.datasetTemplate = cdb.templates.getTemplate('common/views/feed/dataset_item_template');
-    this.loaderTemplatePath = 'common/views/feed/loading';
-    this.emptyTemplatePath = 'common/views/feed/empty';
-    this.errorTemplatePath = 'common/views/feed/error';
+    this.template = cdb.templates.getTemplate('user_feed/template');
+    this.mapTemplate = cdb.templates.getTemplate('user_feed/map_item_template');
+    this.datasetTemplate = cdb.templates.getTemplate('user_feed/dataset_item_template');
+    this.loaderTemplatePath = 'user_feed/loading';
+    this.emptyTemplatePath = 'user_feed/empty';
+    this.errorTemplatePath = 'user_feed/error';
   },
 
   _initModels: function() {
