+++ lib/assets/test/spec/cartodb/common/views/explore/collection.spec.js
@@ -1,7 +1,7 @@
 var $ = require('jquery');
 var cdb = require('cartodb.js');
 cdb.admin = require('cdb.admin');
-var Visualizations = require('../../../../../../javascripts/cartodb/common/views/explore/feed_collection');
+var Visualizations = require('../../../../../../javascripts/cartodb/explore/feed_collection');
 
 describe('common/views/explore/feed_collection', function() {
