+++ lib/assets/test/spec/cartodb/common/views/explore/model.spec.js
@@ -1,7 +1,7 @@
 var $ = require('jquery');
 var cdb = require('cartodb.js');
 cdb.admin = require('cdb.admin');
-var ExploreModel = require('../../../../../../javascripts/cartodb/common/views/explore/model');
+var ExploreModel = require('../../../../../../javascripts/cartodb/explore/model');
 
 describe('common/views/explore/model', function() {
