+++ lib/assets/test/spec/cartodb/common/views/feed/feed.spec.js
@@ -2,8 +2,8 @@ var $ = require('jquery');
 var cdb = require('cartodb.js');
 cdb.admin = require('cdb.admin');
 
-var Feed = require('../../../../../../javascripts/cartodb/common/views/feed/view');
-var Visualizations = require('../../../../../../javascripts/cartodb/common/views/feed/feed_collection');
+var Feed = require('../../../../../../javascripts/cartodb/user_feed/view');
+var Visualizations = require('../../../../../../javascripts/cartodb/user_feed/feed_collection');
 
 describe('common/views/feed/view', function() {
